moves deleteTask to its own file
